A wrapper function to call the Slack API with authentication and pagination.

A wrapper function to call the Slack API with authentication and pagination.

Usage

call_slack_api(
  path,
  ...,
  body = NULL,
  .method = c("GET", "POST"),
  token,
  .verbose = Sys.getenv("SLACKR_VERBOSE", "FALSE"),
  .next_cursor = ""
)

Arguments

path

The API definition path, e.g. /api/auth.test.

...

These arguments must be named and will be added to the API query string.

body

If .method = POST the body gets passed to the API body.

.method

Either "GET" or "POST".

token

Authentication token bearing required scopes.

.verbose

If TRUE, prints httr verbose messages. Useful for debugging.

.next_cursor

The value of the next cursor, when using pagination.

Value

The API response (a named list)