Send the graphics contents of the current device to a Slack channel

dev.slackr sends the graphics contents of the current device to the specified Slack channel.

dev_slackr(
  channels = Sys.getenv("SLACK_CHANNEL"),
  bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN"),
  file = "plot"
)

dev.slackr(
  channels = Sys.getenv("SLACK_CHANNEL"),
  bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN"),
  file = "plot"
)

Arguments

channels
list of channels to post image to

bot_user_oauth_token
the Slack full bot user OAuth token (chr)

file
prefix for filenames (defaults to plot)

Value

httr response object from POST call

Note

You can pass in as_user=TRUE as part of the ... parameters and the Slack API will post the message as your logged-in user account (this will override anything set in username).

References

https://github.com/hrbrmstr/slackr/pull/12/files

See also

Author

Konrad Karczewski [ctb], Bob Rudis [aut]

Examples

if (FALSE) {
slackr_setup()

# base
library(maps)
map("usa")
dev_slackr("#results", file='map')

# base
barplot(VADeaths)
dev_slackr("@jayjacobs")
}