Post a ggplot to a Slack channel

Unlike the slackr_dev() function, this one takes a ggplot object, eliminating the need to have a graphics device (think use in scripts).

ggslackr(
  plot = last_plot(),
  channels = Sys.getenv("SLACK_CHANNEL"),
  scale = 1,
  width = par("din")[1],
  height = par("din")[2],
  units = NULL,
  dpi = 300,
  limitsize = TRUE,
  token = Sys.getenv("SLACK_TOKEN"),
  file = "ggplot",
  initial_comment = NULL,
  thread_ts = NULL,
  title = NULL,
  device = c("png", "eps", "ps", "pdf", "jpeg", "tiff", "bmp", "svg"),
  ...
)

Arguments

plot

ggplot object to save, defaults to last plot displayed.

channels

Comma-separated list of channel names or IDs where the file will be shared.

scale

scaling factor.

width

width (defaults to the width of current plotting window).

height

height (defaults to the height of current plotting window).

units

units for width and height when either one is explicitly specified (in, cm, or mm).

dpi

dpi to use for raster graphics.

limitsize

when TRUE (the default), ggsave will not save images larger than 50x50 inches, to prevent the common error of specifying dimensions in pixels.

token

Authentication token bearing required scopes.

file

Prefix for filenames (defaults to ggplot).

initial_comment

The message text introducing the file in specified channels.

thread_ts

Provide another message's ts value to upload this file as a reply. Never use a reply's ts value; use its parent instead.

title

Title of file.

device

the file extension to use. Options: "png", "eps", "ps", "pdf", "jpeg", "tiff", "bmp", "svg". Default: "png".

...

other arguments passed to graphics device.

Value

httr response object (invisibly)

Examples

if (FALSE) {
slackr_setup()
ggslackr(qplot(mpg, wt, data = mtcars))
}