Save R objects to an RData file on Slack

slackr_save enables you upload R objects (as an R data file) to Slack and (optionally) post them to one or more channels (if channels is not empty).

Usage

slackr_save(
  ...,
  channels = Sys.getenv("SLACK_CHANNEL"),
  file = "slackr",
  token = Sys.getenv("SLACK_TOKEN"),
  initial_comment = NULL,
  title = NULL,
  thread_ts = NULL
)

Arguments

...

objects to store in the R data file.

channels

Comma-separated list of channel names or IDs where the file will be shared.

file

filename (without extension) to use.

token

Authentication token bearing required scopes.

initial_comment

The message text introducing the file in specified channels.

title

Title of file.

thread_ts

Provide another message's ts value to upload this file as a reply. Never use a reply's ts value; use its parent instead.

Value

httr response object from POST call

See also

Examples

if (FALSE) { # \dontrun{
slackr_setup()
slackr_save(mtcars, channels = "#slackr", file = "mtcars")
} # }