Calls the slack API with pagination using cursors.

This loops over fun, extracts the next_cursor from the API response, and injects this into the next loop. At the completion of each loop, the function convert_response_to_tibble() is run with extract as and argument. The results are combined with dplyr::bind_rows()

with_pagination(fun, extract)

Arguments

fun

A function that calls the slack API.

extract

The name of the element to extract from the API response.

Value

A tibble

See also