Output R expressions to a Slack channel/user

Takes an expr, evaluates it and sends the output to a Slack chat destination. Useful for logging, messaging on long compute tasks or general information sharing.

slackr(
  ...,
  channel = Sys.getenv("SLACK_CHANNEL"),
  username = Sys.getenv("SLACK_USERNAME"),
  icon_emoji = Sys.getenv("SLACK_ICON_EMOJI"),
  bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN")
)

Arguments

...
expressions to be sent to Slack

channel
which channel to post the message to (chr)

username
what user should the bot be named as (chr)

icon_emoji
what emoji to use (chr) "" will mean use the default

bot_user_oauth_token
Slack bot user OAuth token

Details

By default, everything but expr will be looked for in a "SLACK_" environment variable. You can override or just specify these values directly instead, but it's probably better to call slackr_setup first.

Note

You need a https://www.slack.com account and will also need to setup an API token https://api.slack.com/ Also, you can pass in as_user=TRUE, the default, as part of the ... parameters and the Slack API will post the message as your logged-in user account (this will override anything set in username). Passing as_user=FALSE, results in the Slack API posting the message as set in username

See also

Examples

if (FALSE) {
slackr_setup()
slackr("iris info", head(iris), str(iris))
}