Sends a message to a slack channel.

Sends a message to a slack channel.

slackr_msg(
  txt = "",
  channel = Sys.getenv("SLACK_CHANNEL"),
  username = Sys.getenv("SLACK_USERNAME"),
  icon_emoji = Sys.getenv("SLACK_ICON_EMOJI"),
  token = Sys.getenv("SLACK_TOKEN"),
  thread_ts = NULL,
  reply_broadcast = FALSE,
  ...
)

Arguments

txt
text message to send to Slack. If a character vector of length > 1 is passed in, they will be combined and separated by newlines.

channel
Channel, private group, or IM channel to send message to. Can be an encoded ID, or a name. See the chat.postMessage endpoint documentation for details.

username
what user should the bot be named as (chr).

icon_emoji
what emoji to use (chr) "" will mean use the default.

token
Authentication token bearing required scopes.

thread_ts
Provide another message's ts value to make this message a reply. Avoid using a reply's ts value; use its parent instead.

reply_broadcast
Used in conjunction with thread_ts and indicates whether reply should be made visible to everyone in the channel or conversation. Defaults to FALSE.

...
other arguments passed to the Slack API chat.postMessage call

Value

the response (invisibly)

Note

You need a https://www.slack.com account and will also need to setup an API token https://api.slack.com/ Also, you can pass in add_user=TRUE as part of the ... parameters and the Slack API will post the message as your logged-in user account (this will override anything set in username)

See also

Examples

if (FALSE) {
slackr_setup()
slackr_msg("Hi")
}